Install work and mod files for import of external modules
