+++ Software/Print_Server_Software/go.work
@@ -1,0 +1,3 @@
+go 1.21.6
+
+use ./printserver
