+++ Software/Print_Server_Software/printserver/go.mod
@@ -1,0 +1,3 @@
+module example.com/printserver
+
+go 1.21.6
